+++ R/api.R
@@ -1,7 +1,7 @@
 #' Returns data for a given flow and key
 #'
 #' @param flow `character(1)` flow to query, 5-8 characters.
-#'   See [bb_metadata()] for available dataflows.
+#'   See [bbk_metadata()] for available dataflows.
 #' @param key `character(1)` key to query.
 #' @param start_period `character(1)` start date of the data. Supported formats:
 #'   - YYYY for annual data (e.g., "2019")
@@ -19,44 +19,37 @@
 #'   start of the series. If `NULL`, no restriction is applied. Default `NULL`.
 #' @param last_n `numeric(1)` number of observations to retrieve from the end
 #'  of the series. If `NULL`, no restriction is applied. Default `NULL`.
-#' @returns A `data.frame()` with the requested data. The columns are:
-#'   \item{date}{The date of the observation}
-#'   \item{value}{The value of the observation}
-#'   \item{key}{The time series key}
-#'   \item{title}{The title of the dataflow}
-#'   \item{frequency}{The frequency of the observation}
-#'   \item{category}{The category of the observation}
-#'   \item{unit}{The unit of the observation}
-#'   \item{unit_multiplier}{The unit multiplier of the observation}
-#' @references <https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-data>
+#' @returns A `data.frame()` with the requested data.
+#' @source <https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-data>
 #' @family data
 #' @export
 #' @examples
 #' \donttest{
 #' # fetch all data for a given flow and key
-#' bb_data("BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A")
-#'
+#' bbk_data("BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A")
+#' # fetch data for multiple keys
+#' bbk_data("BBK01", c("TTA032", "TTA010"))
 #' # specified period (start date-end date) for daily data
-#' bb_data(
+#' bbk_data(
 #'   "BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A",
 #'   start_period = "2020-01-01",
 #'   end_period = "2020-08-01"
 #' )
 #' # or only specify the start date
-#' bb_data(
+#' bbk_data(
 #'   "BBSIS", "D.I.ZAR.ZI.EUR.S1311.B.A604.R10XX.R.A.A._Z._Z.A",
-#'   start_period = "2020-01-01"
+#'   start_period = "2024-04-01"
 #' )
 #' }
-bb_data <- function(flow,
-                    key,
-                    start_period = NULL,
-                    end_period = NULL,
-                    first_n = NULL,
-                    last_n = NULL) {
+bbk_data <- function(flow,
+                     key = NULL,
+                     start_period = NULL,
+                     end_period = NULL,
+                     first_n = NULL,
+                     last_n = NULL) {
   stopifnot(
     is_string(flow), nchar(flow) %in% 5:8,
-    is_string(key),
+    is_character_or_null(key),
     is_string_or_null(start_period),
     is_string_or_null(end_period),
     is_count_or_null(first_n),
@@ -68,6 +61,7 @@ bb_data <- function(flow,
     resource <- sprintf("data/%s", flow)
   } else {
     key <- toupper(key)
+    key <- paste(key, collapse = "+")
     resource <- sprintf("data/%s/%s", flow, key)
   }
   body <- make_request(
@@ -77,42 +71,31 @@ bb_data <- function(flow,
     firstNObservations = first_n,
     lastNObservations = last_n
   )
-  data <- parse_bb_data(body, key)
+  data <- parse_bbk_data(body)
   as_tibble(data)
 }
 
 #' Returns the time serie that is found with the specified time series key
 #'
-#' @param key `character(1)` key to query.
-#' @inherit bb_data references
-#' @returns A `data.frame()` with the requested data. The columns are:
-#'   \item{date}{The date of the observation}
-#'   \item{value}{The value of the observation}
-#'   \item{key}{The time series key}
-#'   \item{title}{The title of the dataflow}
-#'   \item{frequency}{The frequency of the observation}
-#'   \item{category}{The category of the observation}
-#'   \item{unit}{The unit of the observation}
-#'   \item{unit_multiplier}{The unit multiplier of the observation}
-#'   \item{last_update}{The last update date}
-#'   \item{source}{The source of the data}
-#'   \item{comment}{A comment on the data}
+#' @inherit bbk_data source
+#' @inheritParams bbk_data
+#' @inherit bbk_data return
 #' @family data
 #' @export
 #' @examples
 #' \donttest{
-#' bb_series("BBEX3.M.DKK.EUR.BB.AC.A01")
-#' bb_series("BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B")
-#' bb_series("BBBK11.D.TTA000")
+#' bbk_series("BBEX3.M.DKK.EUR.BB.AC.A01")
+#' bbk_series("BBAF3.Q.F41.S121.DE.S1.W0.LE.N._X.B")
+#' bbk_series("BBBK11.D.TTA000")
 #' }
-bb_series <- function(key) {
+bbk_series <- function(key) {
   stopifnot(is_string(key))
   body <- build_request("data/tsIdList", accept = "application/vnd.bbk.data+csv-zip") |>
     req_body_json(key, auto_unbox = FALSE) |>
     req_perform() |>
     resp_body_raw()
 
-  data <- parse_bb_series(body, key)
+  data <- parse_bbk_series(body, key)
   as_tibble(data)
 }
 
@@ -130,17 +113,17 @@ bb_series <- function(key) {
 #' The columns are:
 #'   \item{id}{The id of the metadata}
 #'   \item{name}{The name of the metadata}
-#' @references <https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-metadata>
+#' @source <https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/web-service-interface-metadata>
 #' @family metadata
 #' @export
 #' @examples
 #' \donttest{
-#' bb_metadata("datastructure")
-#' bb_metadata("dataflow", "BBSIS")
-#' bb_metadata("codelist", "CL_BBK_ACIP_ASSET_LIABILITY")
-#' bb_metadata("concept", "CS_BBK_BSPL")
+#' bbk_metadata("datastructure")
+#' bbk_metadata("dataflow", "BBSIS")
+#' bbk_metadata("codelist", "CL_BBK_ACIP_ASSET_LIABILITY")
+#' bbk_metadata("concept", "CS_BBK_BSPL")
 #' }
-bb_metadata <- function(type, id = NULL, lang = c("en", "de")) {
+bbk_metadata <- function(type, id = NULL, lang = c("en", "de")) {
   type <- match.arg(type, c("datastructure", "dataflow", "codelist", "concept"))
   args <- switch(type,
     datastructure = list("datastructure/BBK", "//structure:DataStructure"),
@@ -153,7 +136,7 @@ bb_metadata <- function(type, id = NULL, lang = c("en", "de")) {
   as_tibble(res)
 }
 
-parse_bb_series <- function(body, key) {
+parse_bbk_series <- function(body, key) {
   tmp <- tempfile()
   dir.create(tmp)
   on.exit(unlink(tmp, recursive = TRUE), add = TRUE)
@@ -191,15 +174,11 @@ parse_bb_series <- function(body, key) {
     P1D = "daily"
   )
   res$date <- parse_date(res$date, freq)
-  res$key <- key
-  res$title <- title
-  res$frequency <- freq
-  res$category <- category
-  res$unit <- unit
-  res$unit_multiplier <- unit_mult
-  res$last_update <- last_update
-  res$source <- src
-  res$comment <- comment
+  res <- cbind(
+    res, key, title, freq, category, unit, unit_mult, last_update, comment,
+    source = src
+  )
+  res <- res[union(c("date", "key", "value", "title", "freq"), names(res))]
   res
 }
 
@@ -214,57 +193,62 @@ parse_metadata <- function(x, lang) {
   do.call(rbind, res)
 }
 
-parse_bb_data <- function(body, key) {
-  freq <- body |>
-    xml2::xml_find_first("//generic:Value[@id='BBK_STD_FREQ']") |>
-    xml2::xml_attr("value")
-  freq <- switch(freq,
-    A = "annual",
-    S = "semi-annual",
-    Q = "quarterly",
-    M = "monthly",
-    W = "weekly",
-    D = "daily"
-  )
+parse_bbk_data <- function(body) {
+  series <- body |> xml2::xml_find_all(".//generic:Series")
+  res <- lapply(series, \(x) {
+    series_key <- x |>
+      xml2::xml_find_first(".//generic:SeriesKey") |>
+      xml2::xml_children()
+    nms <- series_key |>
+      xml2::xml_attr("id") |>
+      tolower()
+    series_key <- series_key |>
+      xml2::xml_attr("value") |>
+      stats::setNames(nms) |>
+      as.list()
 
-  key <- body |>
-    xml2::xml_find_first("//generic:Value[@id='BBK_ID']") |>
-    xml2::xml_attr("value")
-  title <- body |>
-    xml2::xml_find_first("//generic:Value[@id='BBK_TITLE_ENG']") |>
-    xml2::xml_attr("value")
-  unit <- body |>
-    xml2::xml_find_first("//generic:Value[@id='UNIT_ENG']") |>
-    xml2::xml_attr("value")
-  unit_mult <- body |>
-    xml2::xml_find_first("//generic:Value[@id='BBK_UNIT_MULT']") |>
-    xml2::xml_attr("value")
-  category <- body |>
-    xml2::xml_find_first("//generic:Value[@id='WEB_CATEGORY']") |>
-    xml2::xml_attr("value")
+    attrs <- x |>
+      xml2::xml_find_first(".//generic:Attributes") |>
+      xml2::xml_children()
+    nms <- attrs |>
+      xml2::xml_attr("id") |>
+      tolower()
+    nms <- replace(nms, nms == "bbk_title_eng", "title")
+    nms <- replace(nms, nms == "bbk_id", "key")
+    attrs <- attrs |>
+      xml2::xml_attr("value") |>
+      stats::setNames(nms) |>
+      as.list()
 
-  entries <- body |> xml2::xml_find_all("//generic:Obs[generic:ObsValue]")
-  date <- entries |>
-    xml2::xml_find_all(".//generic:ObsDimension") |>
-    xml2::xml_attr("value") |>
-    parse_date(freq)
+    data <- c(series_key, attrs)
+    nms <- names(data)
+    nms <- sub("^bbk_(seis_)?", "", nms)
+    nms <- sub("^std_", "", nms)
+    names(data) <- replace(nms, nms == "web_category", "category")
 
-  value <- entries |>
-    xml2::xml_find_all(".//generic:ObsValue") |>
-    xml2::xml_attr("value") |>
-    as.numeric()
+    data$freq <- switch(data$time_format,
+      P1M = "monthly",
+      P3M = "quarterly",
+      P1Y = "annual",
+      P1D = "daily"
+    )
 
-  data <- data.frame(
-    date = date,
-    value = value,
-    key = key,
-    title = title,
-    frequency = freq,
-    category = category,
-    unit = unit,
-    unit_multiplier = unit_mult
-  )
-  data
+    entries <- body |> xml2::xml_find_all("//generic:Obs[generic:ObsValue]")
+    data$date <- entries |>
+      xml2::xml_find_all(".//generic:ObsDimension") |>
+      xml2::xml_attr("value") |>
+      parse_date(data$freq)
+
+    data$value <- entries |>
+      xml2::xml_find_all(".//generic:ObsValue") |>
+      xml2::xml_attr("value") |>
+      as.numeric()
+
+    as.data.frame(data)
+  })
+  res <- do.call(rbind, res)
+  res <- res[union(c("date", "key", "value", "title", "freq"), names(res))]
+  res
 }
 
 parse_date <- function(date, freq) {
@@ -289,7 +273,7 @@ fetch_metadata <- function(resource, xpath, id = NULL, lang = "en") {
   res
 }
 
-bb_error_body <- function(resp) {
+bbk_error_body <- function(resp) {
   body <- resp_body_json(resp)
   message <- body$title
   docs <- "See docs at <https://www.bundesbank.de/en/statistics/time-series-databases/help-for-sdmx-web-service/status-codes/status-codes-855918>" # nolint
@@ -301,7 +285,7 @@ build_request <- function(resource, accept = NULL) {
     req_user_agent("worldbank (https://m-muecke.github.io/worldbank)") |>
     req_headers(`Accept-Language` = "en", accept = accept) |>
     req_url_path_append(resource) |>
-    req_error(body = bb_error_body)
+    req_error(body = bbk_error_body)
 }
 
 make_request <- function(resource, ...) {