+++ NEWS.md
@@ -1,3 +1,10 @@
+# bbk 0.3.0
+
+* Prefix in function name has changed from `bb_` to `bbk_`
+* Allow multiple keys in `bbk_data()`
+* Dynamically parse column names in `bbk_data()`
+* Support for `/data/flowRef` endpoint
+
 # bbk 0.2.0
 
 * Remove missing rows in `bb_series()`